Can switch between fragments
MainActivity has a fragment in it that represents the current screen of
the app. When you tap on a tab on the navigation pane you can switch
between these fragments.

I am using fragments instead of different activities so the app doesn’t
have to recreate the navigation pane and toolbar each time you switch.